Move cross-endpoint block helpers into helper-shared-config.js and keep per-endpoint / singleton-hybrid helpers in helper-session.js.
- Register helper-shared-config in template-engine includedHelpers.
- Classify using SQL GROUP BY: shared queries omit ENDPOINT_IDENTIFIER;  per-endpoint (or singleton-aware hybrid) include it.
- Extend @module JSDoc on both files; regenerate docs/helpers.md.
- Github: ZAP #987
